lec02/hw/job1/requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
requests==2.28.1
black==22.12.0
fastavro==1.9.7
python-dotenv==1.0.0
pytest==7.4.3